+++ pkg/api/unmarshal_test.go
@@ -99,7 +99,7 @@ func TestInvalidStructure(t *testing.T) {
 }
 
 func TestValidPublisMessage(t *testing.T) {
-	intervalNumber := time.Date(2020, 04, 17, 20, 04, 01, 1, time.UTC).Unix() / 600
+	intervalNumber := int32(time.Date(2020, 04, 17, 20, 04, 01, 1, time.UTC).Unix() / 600)
 	json := `{"diagnosisKeys": [
 		  {"key": "ABC", "intervalNumber": %v},
 		  {"key": "DEF", "intervalNumber": %v},
